requirements.txt
torch
torchvision
roma
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
sckit-learn
pyglet<2
huggingface-hub[torch]>=0.22
pillow-heif  # add heif/heic image support
pyrender  # for rendering depths in scannetpp
kapture  # for visloc data loading
kapture-localization
numpy-quaternion
pycolmap  # for pnp
poselib  # for pnp
mast3r
dust3r
colmap
scikit-learn
scipy

